ref(snapshots): Make snapshots toolbar presentational and de-duplicate
Extract the preprod snapshots toolbar into a dedicated snapshotsToolbar.tsx
module and make the components purely presentational. DiffModeToggle now
takes a showSplit prop instead of calling useBreakpoints(), and the toggles
no longer call useOrganization()/trackAnalytics — that wiring moves up into
SnapshotMainContent, which passes breakpoints.sm and fires analytics from
its onChange handlers. snapshotMainContent.tsx imports the shared components
instead of defining its own copies, eliminating the duplication.

Also drop the now-unused export on TRANSPARENT_COLOR in diffImageDisplay.tsx
(still used internally, no longer imported elsewhere after the de-dupe).
